+++ src/suppr-package.c
@@ -16,6 +16,8 @@ extern SEXP C_repeated(SEXP x);
 extern SEXP C_repeated_indices(SEXP x);
 extern SEXP C_repeats(SEXP x);
 extern SEXP C_is_nonfinite(SEXP x);
+extern SEXP C_rm_suffix(SEXP x, SEXP suffix);
+extern SEXP C_rm_prefix(SEXP x, SEXP prefix);
 
 static const R_CallMethodDef callMethods[] = {
     {"C_is_wholenumber", (DL_FUNC)&C_is_wholenumber, 2},
@@ -32,6 +34,8 @@ static const R_CallMethodDef callMethods[] = {
     {"C_repeated_indices", (DL_FUNC)&C_repeated_indices, 1},
     {"C_repeats", (DL_FUNC)&C_repeats, 1},
     {"C_is_nonfinite", (DL_FUNC)&C_is_nonfinite, 1},
+    {"C_rm_suffix", (DL_FUNC)&C_rm_suffix, 2},
+    {"C_rm_prefix", (DL_FUNC)&C_rm_prefix, 2},
     {NULL, NULL, 0}};
 
 void R_init_suppr(DllInfo *dll)