+++ NEWS.md
@@ -1,5 +1,9 @@
 # suppr (development version)
 
+#### New functions
+
+* `rm.prefix()` and `rm.suffix()` to remove a given prefix or suffix from character vector elements.
+
 # suppr 1.0.1
 
 * `whichRepeated()`, `whichNA()`, `whichMin()`/`whichMax()` when `loc = "all"` and `repeats()` now don't attempt to preserve names when no indices or duplicates are found, respectively.